Basic health bar changes
I added a simple health bar system that will get replaced with the drawing. 

For the hp bar, I will have two scoreboards layered upon each other. The scoreboard that's "hidden" will have contrasting colors (if the front is blue, back is red for example). When the fighter takes damage, the front scoreboard will be cropped so the back one is shown which would give the feeling of an hp bar that decreases.